pyproject.toml
[project]
name = "EarthOrbitPlan"
dynamic = ["version"]
authors = [{name = "", email = "kiend.weizman7@gmail.com"}]
license = {text = "BSD-3-Clause"}
description = "Educational Framework for Multimessenger Observing Scenarios"
readme = "README.rst"
requires-python = ">=3.11"
dependencies = [
    "m4opt @ git+https://github.com/m4opt/m4opt@main",
    "pandas",
    "joblib",
    "dask",
    "pre-commit",
    "memory-profiler",
    "python-ligo-lw",
    "igwn-ligolw",
    "memory-profiler",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Educational :: Astronomy",
    "Typing :: Typed",
]

[project.optional-dependencies]
test = [
    "astroplan",
    "networkx",
    "pytest-astropy",
]
docs = [
    "pysiaf",
    "sphinx-astropy[confv2]",
    "sphinxcontrib-bibtex",
    "sphinxcontrib-typer",
    "sphinx-automodapi",
    "sphinx-design",
    "ipykernel",
    "nbsphinx",
    "sphinx-gallery",
    "ipython",
    "ipykernel",
    "jupyter_client",
    "nbconvert",
    "sphinx",
    "numpydoc",
    "graphviz",
    "joblib",
    "dask",
    "ligo.skymap >= 2.4.0",
    "python-ligo-lw",
    "igwn-ligolw",
    "memory-profiler"
]

[project.urls]
Homepage  = "https://EarthOrbitPlan.readthedocs.io"
"Bug Tracker" = "https://github.com/weizmannk/EarthOrbitPlan/issues"
"Change Log" = "https://EarthOrbitPlan.readthedocs.io/en/latest/changes.html"
"Documentation" = "https://EarthOrbitPlan.readthedocs.io/"
"GitHub" = "https://github.com/weizmannk/EarthOrbitPlan"

[tool.setuptools.packages.find]

[tool.setuptools.package-data]
"*" = [
    "py.typed",
    "*.txt",
    "*.ecsv",
]
[tool.pytest.ini_options]
testpaths = [
    "EarthOrbitPlan",
    "docs",
]
norecursedirs = [
    "_build",
]
astropy_header = true
doctest_plus = "enabled"
text_file_format = "rst"
addopts = [
    "--doctest-rst",
    "--ignore=docs/_build",
]
doctest_optionflags = [
    "FLOAT_CMP",
    "IGNORE_EXCEPTION_DETAIL",
]

[tool.coverage.run]
omit = [
    "EarthOrbitPlan/*setup_package*",
    "*/EarthOrbitPlan/*setup_package*",
    "*/EarthOrbitPlan/extern/*",
    "*/EarthOrbitPlan/_version*",
]

[tool.coverage.report]
exclude_also = [
    # Don't complain about packages we have installed
    "except ImportError",
    # Don't complain if tests don't hit assertions
    "raise AssertionError",
    "raise NotImplementedError",
    # Don't complain about script hooks
    "def main\\(.*\\):",
    # Ignore branches that don't pertain to this version of Python
    "pragma: py{ignore_python_version}",
    # Don't complain about IPython completion helper
    "def _ipython_key_completions_",
]


[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "_version.py"

[tool.mypy]
exclude = ["^docs/"]
ignore_missing_imports = true

[tool.codespell]
ignore-words-list = """
    aas,
    sav,
    ser,
    ToO,
    ToOs,
"""
